Feat: [BADA-205] 자기 자신 팔로잉 검증 로직 추가 #198

+++ src/main/java/com/TwoSeaU/BaData/domain/user/service/UserService.java
@@ -109,6 +109,10 @@ public class UserService {
 		final User followingUser = userRepository.findById(userId)
 				.orElseThrow(() -> new GeneralException(UserException.USER_NOT_FOUND));
 
+		if (loginUser.getUsername().equals(followingUser.getUsername())){
+			throw new GeneralException(UserException.CANT_FOLLOW_SELF);
+		}
+
 		return userLikesRepository.findByFollowerUserAndFollowingUser(loginUser, followingUser)
 				.map(userLikes -> {
 					userLikesRepository.delete(userLikes);
